requirements.txt
rich>=13.0.0
psutil>=5.9.0
GPUtil>=1.4.0
nvidia-ml-py>=11.525.0
py-cpuinfo>=9.0.0
pyopencl>=2023.1
numpy>=1.24.0
